+++ CHANGELOG.md
@@ -13,6 +13,11 @@
 - `misc` :
     - Adds library `bits` for manipulating bits.
 #### [&rarrb; v4.2.0](https://github.com/SebastianOberschwendtner/OTOS/releases/tag/v4.2.0)
+- `misc` :
+    - Adds library `units` for custom unit conversion.
+- `battery` :
+    - Refactors *MAX17205* driver to use the new units and register classes.
+    - Refactors *BQ25700* driver to use register classes.
 
 ### Fixed Issues:
 - n/a